Fix menu not cleared outside content area
When closing the menu, sdl_miyoomini_clear_border is not called. When
the menu renders outside of the content area (e.g. with stretched menu
and non 4:3 content), the menu will remain on the framebuffer.

There is an existing check (probably from dingux) that checks if the
menu is closed and `was_in_menu` is set (meaning this is the first frame
where the menu is closed), we clear the border. However, the miyoomini
version does not set `was_in_menu`, so that logic was not triggered.

By saving `was_in_menu`, we can also only set the stOpt.eRotate on the
first frame that the menu opens, and the first frame after the menu is
closed.